Mel Frequency Cepstral Coefficients

Mel Frequency Cepstral Coefficients (MFCCs) are extracted on a frame by frame basis, with one set of coefficients per frame. The extraction is implemented in the frequency domain using the DFT implementation from FFTW.jl. MFCC extraction is provided as one function:

melfcc- Extract a MFCC vector per frame of the input speech signal

SpeechBox.melfcc — Function

melfcc(frames; nfilt=17, ncoeff=13)
melfcc(signal; nfilt=17, ncoeff=13, wind_dur=0.02, win_shift=0.01)
melfcc(x, fs; nfilt=17, ncoeff=13, wind_dur=0.02, win_shift=0.01)

Compute a vector of mel frequency cepstral coefficients for each frame of the input signal

Arguments

frames: a framed_signal object that holds the input signal and details of frames

signal: a speech_waveform object holding the input speech waveform

x: an array holding input speech signal

fs: sampling rate

nfilt: number of triangular mel filters used in the MFCC extraction [default=17]

ncoeff: number of desired MFCC coefficients (must be less than nfilt) including C₀ [default=13]

win_dur: duration of speech frame in secs [default=0.02]

win_shift: time interval between consecutive windows in secs [default=0.01]

SpeechBox.melbankm — Function

melbankm(fs, npts[; nfilt = 17])

Compute a matrix of triangular mel filter responses to multiply with a DFT spectrum given a sampling frequency fs, number of DFT points npts and optionally the number of equally spaced mel filters nfilt

SpeechBox.frq2mel — Function

frq2mel(f)

Convert frequency f from Hz to mel scale

SpeechBox.mel2frq — Function

mel2frq(m)

Convert frequency m from mel scale to Hz