Voice Activity Detection

The vad function takes a framed_signal and generates an array 1s and 0s with each element corresponding to a frame and with a 1 denoting a frame of voiced speech and a 0 denoting unvoiced speech (or an absence of speech). The vad function can make use different voice activity detection algorithms.

VAD algorithms

The Voice Activity Detection algorithms currently implemented are:

Energy Threshold: Assigns every frame with energy greater than a selected threshold as voiced. The threshold is given relative to the maximum frame energy across all input frames.

Energy Fraction: assigns a certain fraction of the frames with the lowest energy as not voiced.

SpeechBox.vad — Function

vad([F, ] sig_frames::framed_signal [; <keyword argument(s)>])

Determine if each frame in the framed_signal object sig_frames corresponds to voiced speech or not using the VAD algorithm specified by F [Default is energy_threshold()]. Keyword arguments are specific to algorithm F.

vad(energy_threshold(), sig_frames::framed_signal [; threshold=0.05])

Estimates the signal 'energy' in each frame (L2 norm computed using the frame_energy function) and assigns every frame with 'energy' greater than threshold times the maximum frame energy as voiced.

vad(energy_fraction(), sig_frames::framed_signal [; unvoiced_fraction=0.2])

Estimates the signal 'energy' in each frame (L2 norm computed using the frame_energy function) and assigns a certain fraction of the frames (specified by unvoiced_fractions) with the lowest energy as not voiced.